Move db helper funcs to own file
